+++ sceneobject.h
@@ -3,9 +3,9 @@
 
 class SceneObject {
 	public:
-		virtual void move(float dx, float dy, float dz) = 0;
-		virtual void rotate(float dxy, float dyz, float dzx) = 0;
-		virtual void resize(float k) = 0;
+		virtual void move(double dx, double dy, double dz) = 0;
+		virtual void rotate(double dxy, double dyz, double dzx) = 0;
+		virtual void resize(double k) = 0;
 };
 
 #endif // SCENEOBJECT_H
